updated tutorial.cs. Now it contains a running tutorial with all features enabled.
updated examples\tutorial with latest changes  (the other examples might not work right now)
updated binding to use latest checkout of core. changed IsValid to IsAttached in c++ and c#
Added group memory read/write functionality in c++ and c#, added unit tests.
added shared group constructor that defaults noCreate and durabillity level using core defaults
modified a few query tests to ensure new query methods gets called from a test
(by letting the dll do the defaulting).
updated release_all script to call the script that updates examples with latest dlls and assemblies
added a few query methods that take columnIndex, that were used in tutorial
Renamed private Transaction method to ExecuteInTransaction in shared group to reduce risk of a misunderstanding
Added support for uses using string[] arrays when calling add with an array of rows for a subtable ,
and alle the column in the subtable are of type string. (because c# would auto default to string[] in that case.
changed unit test to have a call to the SharedGroup constructor that defaults noRead and durabillity level.
Added unit test that exposes the problem we had when the user used compiler-inferred typing of arrays when
sending subtable rows to a table where the rows contains all strings.