changes to make binding work with mono/Windows.
unit tests pass when built with mono 3.10 and run from ximinan studio with mono 3.10 installed
problems were centered around exception handling, as mono does not handle SEH exceptions,
(not even on windows).
also changed ShowInfo to report if we are on Mono, report the version of the CLR running , report of we were built in debug or release mode, and added a call  to ShowInfo to the performance test. changed experimental to show new syntax.
new_group_file now does not throw an exception but instead returns null if an error occoured (that error was very likely a file IO error due to the filename being invalid)