Various issues related to running on .net v 3.5 fixed
Various issues related to running on .net v 4.0 fixed
Now calls Dispose on all exception in constructors of disposable objects, even though it might not be neccessary. We now cannot possibly leak resources due to C# not calling dispose on constructor failures.
Ignored some fxcop warnings that are due to documented bugs in fxcop, some related to 3.5
.net v45, v40 and v35 all pass all unit tests except the usual two know error cases.
No unexpected errors from fxcop in 45 40 and 35, one unexpected errors from resharper in 35, that is most likely a bug in resharper.