work in progress, finishing up tablerow.setsubtable (tests written but need to be finiashed. need testing on .35 .40 and .45 - some of the code is platform specific)
work in progress, performance tests.
validation in TableOrView moved from public methods to methods that validate column index. now table.isvalid is checked just before the column index is looked up or validated. If an operation does not involve a column index, validation is still done in the public method.