Fix so that single threaded applications that do not use using, do not crash anymore.
multi threaded fix to be done.
utilizees a list of handles in user thread that is filled concurrenlty from the finalizer thread
might leak unbind and delete calls on program exit, but will not call core concurrently in a
single threaded app. (hooking into the application shutdown events makes an app threaded)